project/requirements.txt
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
peft>=0.7.0
wandb>=0.15.0
python-dotenv>=1.0.0
pytest>=7.4.0
numpy>=1.24.0
unsloth>=0.3.0 